[backend] Move onGameJoin start logic to separate handler
