+++ webpack/handlers/index.js
@@ -1,3 +1,4 @@
+exports.onActionHandler = require('./onAction');
 exports.onGameJoinHandler = require('./onGameJoin');
 exports.onGameLeaveHandler = require('./onGameLeave');
-exports.onActionHandler = require('./onAction');
+exports.onGameStartHandler = require('./onGameStart');
