+++ webpack/server.js
@@ -14,9 +14,10 @@ const compiler = webpack(config);
 const { bindSocket } = require('./utils');
 const game = require('./routes/api/game');
 const {
+  onActionHandler,
   onGameJoinHandler,
   onGameLeaveHandler,
-  onActionHandler,
+  onGameStartHandler,
 } = require('./handlers');
 
 // - Routes -------------------------------------------------------------------/
@@ -40,6 +41,7 @@ server.listen(config.port, 'localhost', (error) => {
 
     socket.on('gameJoin', bindSocketCurried(onGameJoinHandler));
     socket.on('gameLeave', bindSocketCurried(onGameLeaveHandler));
+    socket.on('gameStart', bindSocketCurried(onGameStartHandler));
     socket.on('action', bindSocketCurried(onActionHandler));
   });
